--HG--
branch : spec_SFEperl-scalar-util

+++ specs/SFEperl-scalar-util.spec
@@ -9,13 +9,16 @@
 %include Solaris.inc
 %include packagenamemacros.inc
 
-%define tarball_version 1.32
+%define build584 1
+%define build512 1
+
+%define tarball_version 1.42
 %define tarball_name    Scalar-Util
 
 Name:		SFEperl-scalar-util
 IPS_package_name: library/perl-5/scalar-util
-Version:	1.32
-IPS_component_version: 1.32
+Version:	%{tarball_version}
+IPS_component_version: %{tarball_version}
 Summary:	Scalar utilities (dualvar reftype etc)
 License:	Artistic
 Distribution:   OpenSolaris
@@ -25,9 +28,6 @@ SUNW_Basedir:	%{_basedir}
 SUNW_Copyright: %{name}.copyright
 Source0:	http://search.cpan.org/CPAN/authors/id/P/PE/PEVANS/Scalar-List-Utils-%{tarball_version}.tar.gz
 
-BuildRequires:	runtime/perl-584
-BuildRequires:	runtime/perl-512
-
 Meta(info.maintainer):          roboporter by pkglabo.justplayer.com <pkgadmin@justplayer.com>
 Meta(info.upstream):            Graham Barr <gbarr@pobox.com>
 Meta(info.upstream_url):        http://search.cpan.org/~gbarr/%{tarball_name}-%{tarball_version}
@@ -36,42 +36,53 @@ Meta(info.classification):	org.opensolaris.category.2008:Development/Perl
 %description
 Scalar utilities (dualvar reftype etc)
 
+%if %{build584}
 %package 584
 IPS_package_name: library/perl-5/scalar-util-584
 Summary: Scalar utilities (dualvar reftype etc) for perl-584
 BuildRequires:	runtime/perl-584
 Requires:	runtime/perl-584
+%endif
 
+%if %{build512}
 %package 512
 IPS_package_name: library/perl-5/scalar-util-512
 Summary: Scalar utilities (dualvar reftype etc) for perl-512
 BuildRequires:	runtime/perl-512
 Requires:	runtime/perl-512
-
+%endif
 
 %prep
+if [ -d $RPM_BUILD_ROOT ]
+then
+    rm -rf $RPM_BUILD_ROOT
+fi
+
 %setup -q -n Scalar-List-Utils-%{tarball_version}
 
 %build
+
+%if %{build584}
 export PERL5LIB=/usr/perl5/vendor_perl/5.8.4
 /usr/perl5/5.8.4/bin/perl Makefile.PL PREFIX=%{_prefix} \
   DESTDIR=$RPM_BUILD_ROOT \
   LIB=/usr/perl5/vendor_perl/5.8.4
 make
 make test
-
-rm -rf $RPM_BUILD_ROOT
 make pure_install
+%endif
 
+%if %{build512}
 export PERL5LIB=/usr/perl5/vendor_perl/5.12
 /usr/perl5/5.12/bin/perl Makefile.PL PREFIX=%{_prefix} \
   DESTDIR=$RPM_BUILD_ROOT \
   LIB=/usr/perl5/vendor_perl/5.12
 make
 make test
+make pure_install
+%endif
 
 %install
-make pure_install
 mkdir -p $RPM_BUILD_ROOT%{_datadir}
 mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
 mv $RPM_BUILD_ROOT%{_datadir}/man/man3 $RPM_BUILD_ROOT%{_datadir}/man/man3perl
@@ -87,15 +98,23 @@ rm -rf $RPM_BUILD_ROOT
 #%attr(0755,root,bin) %dir %{_bindir}
 #%{_bindir}/*
 
+%if %{build584}
 %files 584
 %defattr (-, root, bin)
 %{_prefix}/perl5/vendor_perl/5.8.4
+%endif
 
+%if %{build512}
 %files 512
 %defattr (-, root, bin)
 %{_prefix}/perl5/vendor_perl/5.12
+%endif
 
 %changelog
+* Tue May 26 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 1.42
+* Tue Nov 05 2013 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 1.35
 * Mon Sep 09 2013 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 1.32
 * Thu Jun 14 2012 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
